Removed Go wrapper and restructured/renamed folders
