--- lsl/go.mod
@@ -1,3 +1,0 @@
-module lsl
-
-go 1.23
