+++ rust_simulator/main.rs
